Add pre-made/set-up image files
Added the ABA images folder which contains subfolders labled with
gene names. In these subfolders are the downloaded and cropped
images from the ABA.
Added the expression maps folder which contains subfolders labled
with gene names. In these subfolders, are the already-made gene
expression maps.
Added the template images folder. The images inside this folder
are used as a backdrop when visualizing the gene expression data.
Updated my notes text file.