Edit GUI and create module for detecting/filtering expression
The current main features of the GUI are the image viewer, the next
button, the previous button, the current image number tracker, the
listbox, and the okay button.
The MainGUI module also compares the expression maps of all the genes
selected in the listbox at the current image in the sequence. This
comparison is done on-the-fly as the user selects which combinations
of gene expression data they want to see.

The module filter_expression decides if there are new, un-processed
ISH images. If there are, it will detect the expression from the
ISH image then filter the expression based on global constants for
window/neighborhood size and minimun expression density.