Add content to the developer's Reference and User Guide documents
The user guide is aimed for the use of investigators using the
program. It includes topics of outside software requirements as well
as instructions on starting the program, using the program, and the
current manual pre-processing work that needs to be done for adding
new gene expression data.

The developer’s reference document is aimed for the user of deveopers
who wish to expand the program. It includes the topics of outside
software requirements, and a brief overview each module. The overview
of each module explains key fucnctions and known bugs.