Add temp folder
This folder is needed as a location to save a temp expression map
image. A empty txt file is inside as a placeholder so that git sees
that the folder should not be ignored. This folder should be deleted
after the issue of converting matplotlib AxesImage to PIL image is
fixed.